automatically generate _my_tags/<tag>.md
added a new bash script to inspect posts and generate appropriate
entries in _my_tags/<tag>.md

should respect the settings for "_my_tags" in config.yml, but it
doesn't at this time...just hard coded to _my_tags

tag_scraper will not modify existing files...what if the user decides
to change the "slug"...I don't want to clobber that.

tag_scraper will not delete files that are no longer
needed...deletion needs to be a manual thing.

integrated tag_scraper into the package.json so that it runs on `npm
run local` and `npm run publish`